+++ openpose/src/main/java9/module-info.java
@@ -1,0 +1,10 @@
+module org.bytedeco.openpose {
+  requires transitive org.bytedeco.javacpp;
+  requires transitive org.bytedeco.caffe;
+  requires transitive org.bytedeco.hdf5;
+  requires transitive org.bytedeco.opencv;
+  requires transitive org.bytedeco.openblas;
+  exports org.bytedeco.openpose.global;
+  exports org.bytedeco.openpose.presets;
+  exports org.bytedeco.openpose;
+}